pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "receipt-ocr-pipeline"
version = "1.0.0"
description = "Local, scriptable pipeline to process receipt scans into structured data and submission-ready PDFs"
readme = "README.md"
authors = [
    {name = "Receipt OCR Pipeline Contributors"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "PyMuPDF>=1.23.0",
    "reportlab>=4.0.0",
    "pypdf>=3.17.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
# Additional LLM providers (OpenAI is included by default)
anthropic = [
    "anthropic>=0.18.0",
]
all-llm = [
    "anthropic>=0.18.0",
]

[project.scripts]
receipt-ocr = "receipt_ocr_pipeline.cli.main:main"

[project.urls]
Homepage = "https://github.com/blotzy/receipt-ocr-pipeline"
Issues = "https://github.com/blotzy/receipt-ocr-pipeline/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["receipt_ocr_pipeline*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
